Crate: move filter rasterization to rendering
Pixel-application and rasterization (blur, drop-shadow, backdrop,
batched pixel filters) move from `style::properties::filter` to a new
`rendering::filter` module. The `Filter` value type, `FromCss`/`ToCss`,
`Animatable`, and filter analysis (`categorize`/`transfer_table`/LUT
builders) stay in `style`, which no longer depends on `rendering`.
